+++ app/models/access_token.rb
@@ -3,5 +3,10 @@
 class AccessToken < ApplicationRecord
   include Uuidable
 
+  FORMAT_BY_PROVIDER = {
+    Providerable::GITHUB => 'github_pat_',
+    Providerable::GITLAB => 'glpat-'
+  }.freeze
+
   belongs_to :tokenable, polymorphic: true
 end